Moved message into the hud (#94)
* Removed popularity from Player.ts

* Added isSpaceJustDown() function

* Working on getting HUD to cleanly slide up/down

* LINT AND FORMAT

* commented out unused stuff

* moving into the hud, yo

* updated reference in main.ts

* updated hud_renderer reference to message.ts in the hud
